+++ src/main/java/com/gregtechceu/gtceu/api/registry/registrate/builder/package-info.java
@@ -1,0 +1,8 @@
+@NotNullByDefault
+@NullMarked
+package com.gregtechceu.gtceu.api.registry.registrate.builder;
+
+import org.jetbrains.annotations.NotNullByDefault;
+import org.jspecify.annotations.NullMarked;
+
+import javax.annotation.ParametersAreNonnullByDefault;
